@@ -5,16 +5,16 @@
     "@salesforce-ux/design-system": "^2.19",
     "@salesforce/design-system-react": "^0.10.48",
     "@types/har-format": "^1.2.8",
-    "@types/jest": "^24.0.13",
+    "@types/jest": "^27.5.2",
     "@types/jsforce": "^1.9.43",
-    "@types/react": "^16.8.19",
-    "@types/react-dom": "^16.8.4",
+    "@types/react": "^18.0.19",
+    "@types/react-dom": "^18.0.6",
     "@types/webextension-polyfill": "^0.9.0",
     "jsforce": "^1.11.0",
     "logviewer": "https://gitpkg.now.sh/swisscat/debug-log-analyzer/log-viewer?main",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-scripts": "3.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "^3.4.4",
     "typescript": "^3.7.5",
     "webextension-polyfill": "^0.10.0"
   },
@@ -31,9 +31,6 @@
     "serve": "serve -s build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "clean": "rm -rf build",
-    "prebuild": "npm run clean",
-    "prewatch": "npm run clean",
     "postinstall": "postinstall",
     "zip": "bestzip extension.zip build/*"
   },
