@@ -1,13 +1,14 @@
 import { IconSettings } from "@salesforce/design-system-react";
 import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOM from "react-dom/client";
 import App from "./App";
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById("root") as HTMLElement);
+
+root.render(
   <React.StrictMode>
     <IconSettings iconPath="/lib/icons">
       <App />
     </IconSettings>
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
